src/test/java/jpa/querydsl/repository/MemberJpaRepositoryTest.java
package jpa.querydsl.repository;

import jakarta.persistence.EntityManager;
import jpa.querydsl.Entity.Member;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;
import java.util.Optional;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
@Transactional
class MemberJpaRepositoryTest {

    @Autowired EntityManager em;
    @Autowired MemberJpaRepository memberJpaRepository;

    @Test
    public void basicTest(){
        Member member = new Member("member1", 10);
        memberJpaRepository.save(member);

        Optional<Member> findMember = memberJpaRepository.findById(member.getId());

        assertThat(findMember.get().getUsername()).isEqualTo("member1");
        assertThat(member).isEqualTo(findMember);

        List<Member> findMembers = memberJpaRepository.findAll();


    }

}